[627] Ядро / Разное: Перевод сохранения\загрызки\апдейта карренси на PS (Prepared Statement) Хранение данных в CurrencyCap по 2 байта, а не 4. Тупо в базе держать по 2 байта, а в ядре 4, ибо это может привести к непредвиденным последствиям Обьеденение BgConquestRatingCalculator и ConquestRatingCalculator

+++ src/server/game/Entities/Player/CurrencyMgr.h
@@ -12,10 +12,10 @@
 
 struct CurrencyCap
 {
-    uint32 highestArenaRating;
-    uint32 highestRBgRating;
-    uint32 currentArenaCap;
-    uint32 currentRBgCap;
+    uint16 highestArenaRating;
+    uint16 highestRBgRating;
+    uint16 currentArenaCap;
+    uint16 currentRBgCap;
     uint8 requireReset;
 };
 
@@ -38,7 +38,7 @@ public:
     }
 
     void AddCurrencyCapData(uint32 lowGuid, uint16 arenaRating = 0, uint16 RBgRating = 0, uint16 arenaCap = DEFAULT_ARENA_CAP, uint16 rbgCap = 0, uint8 reset = 0);
-    void CalculatingCurrencyCap(uint32 &rating, bool ratedBattleground = false);
+    void CalculatingCurrencyCap(uint16 &rating, bool ratedBattleground = false);
     
     CurrencyCap* getCurrencyCapData(uint32 lowGuid);
